Bullets - Work in progress
In this version of the code, we have a ManLEE sprite loaded, and bullets
are somewhat working but still need a lot of improvement. The current
issues are: Bullets will sometimes shoot in odd directions, and at
different speeds - sometimes the bullets will not even fire!
S to shoot, arrow keys to move.