Bullet Collision - Work in progress
Hi guys, so I've now got the enemies to move around by themselves, at
the moment they all just chase the player, but we'll get some better
functionality going as time goes on. This is more of just a source
control update. The bullets now have collision detection, so when they
hit an enemy, the enemy is respawned elsewhere. Again, a work in
progress, but it's progress! I've also added Player collision with
enemies, which again makes the enemies spawn in random places when they
collide with the player.