Bullets working as intended, enemies around 75%
So yeah guys the bullets are now finished, we just need to load the
bullet sprite sheet and configure it so the bullet faces in the
direction it was shot. We also need to try and get the machine gun
function working correctly. The enemies just need specific actions now,
which I'm sure we'll implement pretty soon!