Code clean-up
Hi guys, just done some cleaning up of the code. Removed obsolete
variables and classes, tidied up the textures folder as there were a few
images that we didn't need anymore. Game is still the same as the other
version, but now the could is easier to read! - Nick (P.S. all commits
below this are from Nick)